requirements.txt
numpy
#torch>=1.7.0
tqdm
pyyaml
numba>=0.39.0
#torch_scatter>=1.3.1
Cython
scipy
dropblock
open3d
onnx

notebook==6.4.10
traitlets==5.9.0
